Make UnionAttribute a direct reference
- Splits the library into 2 projects, the attribute and the generator
- Moves common project items to the Directory.Build.props
- Updates Benchmarks to remove covering code that is no longer generated
- Updates tests to work with new project layout
- Adds SDK reference to PackAsAnalyzer
- Updates namespaces
