+++ benchmark/Dunet.Benchmark/Dunet.Benchmark.csproj
@@ -12,7 +12,8 @@
     </ItemGroup>
 
     <ItemGroup>
-      <ProjectReference Include="..\..\src\Dunet.csproj" />
+      <ProjectReference Include="..\..\src\Dunet\Dunet.csproj" />
+      <ProjectReference Include="..\..\src\Dunet.Generator\Dunet.Generator.csproj" />
     </ItemGroup>
 
 </Project>
