+++ src/Dunet.Generator/UnionAttributeGeneration/UnionAttributeSource.cs
@@ -1,0 +1,8 @@
+﻿namespace Dunet.Generator.UnionAttributeGeneration;
+
+internal static class UnionAttributeSource
+{
+    public const string Namespace = "Dunet";
+    public const string Name = "UnionAttribute";
+    public const string FullyQualifiedName = $"{Namespace}.{Name}";
+}
