--- src/Dunet.csproj
@@ -1,52 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-    <PropertyGroup>
-        <TargetFramework>netstandard2.0</TargetFramework>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <Nullable>enable</Nullable>
-        <LangVersion>12.0</LangVersion>
-        <IsPublishable>True</IsPublishable>
-        <IncludeBuildOutput>false</IncludeBuildOutput>
-        <GeneratePackageOnBuild>True</GeneratePackageOnBuild>
-        <Authors>domn1995</Authors>
-        <Company />
-        <Description>A simple source generator for discriminated unions in C#.</Description>
-        <PackageProjectUrl>https://github.com/domn1995/dunet</PackageProjectUrl>
-        <PackageReadmeFile>Readme.md</PackageReadmeFile>
-        <RepositoryUrl>https://github.com/domn1995/dunet</RepositoryUrl>
-        <PackageTags>source; generator; discriminated; union; functional; tagged;</PackageTags>
-        <AssemblyVersion>1.10.0</AssemblyVersion>
-        <FileVersion>1.10.0</FileVersion>
-        <PackageLicenseExpression>MIT</PackageLicenseExpression>
-        <PackageReleaseNotes>https://github.com/domn1995/dunet/releases</PackageReleaseNotes>
-        <RepositoryType>git</RepositoryType>
-        <PackageIcon>favicon.png</PackageIcon>
-        <SignAssembly>False</SignAssembly>
-        <Version>1.10.0</Version>
-        <NeutralLanguage>en</NeutralLanguage>
-        <DevelopmentDependency>true</DevelopmentDependency>
-        <NoWarn>$(NoWarn);NU5128</NoWarn>
-        <EnforceExtendedAnalyzerRules>true</EnforceExtendedAnalyzerRules>
-    </PropertyGroup>
-
-    <ItemGroup>
-        <None Include="$(OutputPath)/$(AssemblyName).dll" Pack="true" PackagePath="analyzers/dotnet/cs" Visible="false" />
-    </ItemGroup>
-
-    <ItemGroup>
-        <None Include="../Readme.md" Pack="true" PackagePath="" Visible="false" />
-        <None Include="../favicon.png" Pack="true" PackagePath="" Visible="false" />
-    </ItemGroup>
-
-    <ItemGroup>
-        <PackageReference Include="Microsoft.CodeAnalysis.Analyzers" Version="3.3.4">
-            <PrivateAssets>all</PrivateAssets>
-            <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-        </PackageReference>
-        <PackageReference Include="Microsoft.CodeAnalysis.CSharp" Version="4.8.0" />
-        <PackageReference Include="PolySharp" Version="1.14.1">
-            <PrivateAssets>all</PrivateAssets>
-            <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-        </PackageReference>
-    </ItemGroup>
-
-</Project>
