+++ src/Dunet/Dunet.csproj
@@ -1,0 +1,15 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <TargetFrameworks>netstandard2.0;net8.0</TargetFrameworks>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+        <LangVersion>12.0</LangVersion>
+        <IsPublishable>True</IsPublishable>
+        <GeneratePackageOnBuild>True</GeneratePackageOnBuild>
+    </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Dunet.Generator\Dunet.Generator.csproj" PackAsAnalyzer="true" PrivateAssets="all" />
+  </ItemGroup>
+</Project>
