+++ src/Dunet/UnionAttribute.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace Dunet;
+
+/// <summary>
+/// Enables dunet union source generation for the decorated partial record.
+/// </summary>
+[AttributeUsage(AttributeTargets.Class, Inherited = false, AllowMultiple = false)]
+public sealed class UnionAttribute : Attribute {}
