+++ test/Dunet.Test.csproj
@@ -22,7 +22,8 @@
     </ItemGroup>
 
     <ItemGroup>
-        <ProjectReference Include="..\src\Dunet.csproj" />
+        <ProjectReference Include="..\src\Dunet\Dunet.csproj" />
+        <ProjectReference Include="..\src\Dunet.Generator\Dunet.Generator.csproj" />
     </ItemGroup>
 
 </Project>
